moving methods to functions
